Consolidate experimental tests into main test suite
- Move experimental tests to appropriate locations in main test structure:
  - test_datatable.py → tests/unit/test_datatable_inheritance.py
  - test_llm.py → tests/unit/llms/test_instructor_factory.py
  - test_metric_base.py → integrated into tests/unit/test_metric.py
- Integrate experimental mock fixtures into main conftest.py
- Update Makefile to remove tests/experimental references
- Remove obsolete process_experimental_notebooks.py script
- Remove tests/experimental/ directory

All tests already used direct ragas imports, simplifying consolidation.
Preserves all test functionality while improving organization.
