+++ tests/unit/llms/test_instructor_factory.py
